requirements.txt
streamlit 
pandas 
joblib
matplotlib.pyplot 
seaborn
scikit-learn
plotly